refactor: folder structure to follow manifest.json
