--- index.js
@@ -1,1 +1,0 @@
-// Adding temporary placeholder file so eslint has something to work on and not barf
