+++ src/background/index.js
@@ -1,4 +1,4 @@
-import RequestBlocker from './lib/request_blocker.js';
+import RequestBlocker from './background/request_blocker.js';
 
 console.log('Extension is ready to start blocking!');
 RequestBlocker.startMonitoring();
