+++ src/background/list_populator.js
