+++ src/background/model/list.js
