+++ src/background/request_blocker.js
