+++ src/background/request_listener.js
