+++ src/manifest.json
@@ -8,14 +8,14 @@
   },
   "permissions": ["webRequest", "webRequestBlocking", "<all_urls>"],
   "background": {
-    "page": "background_script.html"
+    "page": "background/index.html"
   },
   "content_scripts": [
     {
       "matches": ["<all_urls>"],
       "all_frames": true,
       "run_at": "document_start",
-      "js": ["content_script.js"]
+      "js": ["content/index.js"]
     }
   ],
   "browser_action": {
